[VIDEOIO] as per Davis' suggestion, just put the contents of video_capture.h in video_io.h
